refactor: move Stylesheet to include root directory
